Image Nonexistent
I think my copy of eclipse is corrupted or something
It's giving me an image error for a perfectly okay line (I even replaced the image, and made sure the extension and name were correct)
I even commented and eventually cut the line of code it didn't like and ran it--it acted as if it were there (I pasted it back in afterwards)
Even after restarting Eclipse, it gave the same error
It also acted as if i had never imported JavaZoom, and after i restarted JavaZoom was gone
______________
in other news, I added some OTHER code to allow the fireplace image to appear if one of the seven or so randomly picked random colors is picked (say that seven times fast)!
Song is [besides for this] up and running!